Converted to use a ResourceBundle from a property file in the jar.  This replaces ~/query.prop.
The servers are coded as ip-address (as opposed to CNAMES) in the property file.  I don't know if the code copes with using DNS.  This is a possible deployment issue.  It may be possible to pull the property file from a URL and configure the code that way.  Depends on deployment environment.